src/backend/compare-image/requirements.txt
functions-framework==3.*
google-cloud-storage
Pillow
scikit-image
numpy
requests
google-auth
